backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
xai-sdk
resend